split ApiUtils in two
this makes the NotificationController simpler
