+++ app/lib/CORSable.scala
@@ -1,7 +1,7 @@
 package lib
 
-import config.Config.defaultExecutionContext
 import play.api.mvc.{Action, BodyParser, Request, Result}
+import scala.concurrent.ExecutionContext.Implicits.global
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -18,5 +18,5 @@ case class CORSable[A](allowedOrigins: Set[String])(action: Action[A]) extends A
 
   lazy val parser: BodyParser[A] = action.parser
 
-  override def executionContext: ExecutionContext = config.Config.defaultExecutionContext
+  override def executionContext: ExecutionContext = scala.concurrent.ExecutionContext.Implicits.global
 }
